Squashed commit of the following:
commit 51733aac6250f17c4a327481af6285df4b0c15d3
Author: javics2002 <49459590+javics2002@users.noreply.github.com>
Date:   Sat Feb 19 10:20:54 2022 +0100

    Antes de la gran transcendencia

    La gran transcendencia es borrar el proyecto (usa la plantilla de tpv1) y meter nuestro codigo en la plantilla de tpv2

commit fa2b7692988e62555b038766374107cd65071625
Author: javics2002 <49459590+javics2002@users.noreply.github.com>
Date:   Fri Feb 18 09:08:23 2022 +0100

    1 error menos lets go

commit 31fc8f3a1574d2a16d82fed75aae6ce030ffa4c7
Author: javics2002 <49459590+javics2002@users.noreply.github.com>
Date:   Thu Feb 17 09:58:11 2022 +0100

    Me cargo algunos errores de link

commit 0ed7c8fa5ee865e5d1d8ca14d169441340886f36
Author: javics2002 <49459590+javics2002@users.noreply.github.com>
Date:   Thu Feb 17 08:54:05 2022 +0100

    Cambio includes y librerias

commit 258ee0569995e8c99b4b65fdfc3d556df4dd78b5
Author: javics2002 <49459590+javics2002@users.noreply.github.com>
Date:   Sat Feb 12 12:19:52 2022 +0100

    Intento de renderizar fondo

commit b65880cc1102a9cf809197e71d8871afb10c127c
Author: javics2002 <49459590+javics2002@users.noreply.github.com>
Date:   Fri Feb 11 12:42:33 2022 +0100

    Load Tilemap (no compila)

+++ GameObjects/GameObject.cpp
@@ -19,7 +19,7 @@ void GameObject::setPosition(Vector2D<double> v)
 	pos = v;
 }
 
-void GameObject::setDimension(double width, double height)
+void GameObject::setDimension(int width, int height)
 {
 	w = width;
 	h = height;
